backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
exceptiongroup==1.3.1
fastapi==0.129.0
greenlet==3.3.1
h11==0.16.0
httptools==0.7.1
idna==3.11
PyJWT==2.10.1
cryptography==46.0.3
psycopg==3.3.3
psycopg-binary==3.3.3
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.20
python-dotenv==1.2.1
PyYAML==6.0.3
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
pytest==8.4.2
